better num formatting (sco notation, locale fix); fix some ints that should have been longs

+++ core/src/main/java/io/github/celosia/sys/battle/Stats.java
@@ -163,7 +163,7 @@ public class Stats {
 		};
 	}
 
-	public Stats getRealStats(int lvl) {
+	public Stats getRealStats(long lvl) {
 		return new Stats((hp + (hp / 2) * lvl) * STAT_MULT_VISIBLE * STAT_MULT_HIDDEN,
 				(str + (str / 2) * lvl) * STAT_MULT_VISIBLE * STAT_MULT_HIDDEN,
 				(mag + (mag / 2) * lvl) * STAT_MULT_VISIBLE * STAT_MULT_HIDDEN,
